refactor(composition): shrink deployment-mode branching ratchet 5->3 (§4.4/#6274 Track 1)
The `reborn_deployment_mode_branching_ratchet` allowlist tracks every file in
`ironclaw_reborn_composition` still permitted to name a deployment-mode profile
variant; the §4.4 target is `{deployment.rs}` (the one sanctioned profile->config
edge). Two of the five entries were kept alive only by inline `#[cfg(test)]`
fixtures — retire them by extracting the test modules to sibling `tests.rs`
files (which the ratchet scanner excludes):

- `webui/facade.rs` -> `webui/facade/tests.rs`: the `RebornCompositionProfile`
  literals were entirely in the trailing `#[cfg(test)] mod tests`.
- `factory.rs` -> `factory/tests.rs`: same — all profile literals lived in the
  inline test module (production factory.rs names no variant; it reads
  `DeploymentConfig::storage_shape()`). Matches the crate's existing
  `mod auth_tests;` / `mod local_dev_host_tests;` sibling-file pattern. Shrinks
  factory.rs 8566->5599.

Allowlist now `{deployment.rs (target), local_runtime_profile.rs, readiness.rs}`:

- `local_runtime_profile.rs` — investigated for folding into `deployment.rs`; it
  keeps one production branch selecting the hosted-single-tenant-volume build
  input, whose `libsql`-feature requirement no `DeploymentConfig` axis yet
  captures. Cleanly retiring it needs a new config axis, so it's deferred to
  Phase 4 / §5.11 (the `build_runtime(cfg)` collapse), documented on the entry.
- `readiness.rs` — profile is an operator-facing wire **label**
  (`RebornReadinessDiagnostic::profile`), not a branch; stays until that field
  is reshaped.

Behavior-preserving: the extractions are pure moves (super-relative paths
unchanged); `factory/tests.rs` carries an `arch-exempt: large_file` note. No
production code changed.

[skip-regression-check] test-module relocation + ratchet-allowlist trim; the
architecture ratchet itself is the pinning test.
